ref: Move `honeycombTrials` to the trials folder
